better way of handling compiler flag tests; also pass -mt to compiler on Solaris. gcc fixes
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@17775 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ willow/src/willow/willow.cc
@@ -520,7 +520,7 @@ diobuf::diobuf(size_t size)
 	}
 	char	path[PATH_MAX];
 	snprintf(path, sizeof(path), "/dev/shm/willow.diobuf.%d.%d.%d",
-		getpid(), (int) pthread_self(), rand());
+		(int) getpid(), (int) pthread_self(), rand());
 		if ((_fd = open(path, O_CREAT | O_EXCL | O_RDWR, 0600)) == -1) {
 		wlog(WLOG_WARNING, format("opening diobuf %s: %e") % path);
 		_buf = new char[size];
